Print better handles layers not visible
This addresses a javascript error when printing directions. When a layer is not visible at the current zoom level, its value in an array of visible layers is undefined, causing an error when we retrieve layers in the print function. This change tests for the visible layer and, if missing, returns a simpler layer configuration.