Remove -TLINK from Snowplow app ID
After further discussion with Brendan of the GDX team, we agreed to reuse the same app ID in the tracker across hosted SMK apps. They distinguish sites by the URL of the apps the tracker sends requests from, so the app ID does not need to be unique per hosted SMK app.