Prepare layer MRN High Zoom
The Major Road Network High Zoom layer does not seem to appear in the app at the highest zoom level, where it seems we transition from the Major Road Network layer. It looks like the MRN High Zoom layer is not loading.

This change updates the index page to call prepareLayer() for the MRN High Zoom Layer in addition to the existing call to prepare the MRN layer.